Fix camera-relative rendering eliminates label/trail jitter near small bodies
Float32 world-space positions lose ~27 km of precision at Mars distance,
causing labels and trails to snap visibly when the camera moves near moons
or dwarf planets. Fix by storing trail positions as double, subtracting the
camera origin in double before the float32 cast, and passing a translation-
free VP matrix (proj × view_rot) to the trail and label shaders.

+++ src/body.h
@@ -26,7 +26,7 @@ typedef struct {
     double trail_accum;        /* accumulator toward next sample            */
     int    trail_head;
     int    trail_count;
-    float  trail[TRAIL_LEN][3];
+    double trail[TRAIL_LEN][3]; /* AU, camera-relative subtraction done at render time */
 } Body;
 
 extern Body g_bodies[MAX_BODIES];
